requirements.txt
flask
elasticsearch
sentence-transformers
faiss-cpu
numpy
wikipedia